Get user permission for installation
If script is run on a production (or live) environment, packages should
not be installed unless proper change procedures or followed so as not
to increase the attack surface or cause system outage.

Mades changes
for the script to ask user permission before it tries installing
sysv-rc-conf, bc and finger packages.

Fixed few spelling mistakes.